2022 day 14
I don't like sand. It's coarse and rough and irritating and it gets
everywhere.

I didn't like todays problem, mainly because of my own choices. Of
course a `while True` loop won't exit without an break 🙃🙃🙃
Not to mention my off-by-one error when finding the walls 🙃🙃🙃

Had to check other solutions to see what I was doing wrong. Part 2 is
brute-forced by actually adding the 'infinite' wall. It's not pretty
but it works.

Overall mood: 🙃