2022 day 21
Part one I started doing a recursive parse so that I could call `eval`
but then I realised that dict lookups were much easier.

Part 2, still not wanting to do a recursive parse I decided to make the
sums so I could `eval` the completed expression and then paste it into
WolframAlpha. Unfortunately the expression was too long so I did a
binary search. This search is trash, on the test input using less than
as the comparison gets the right answer but not on my real input.
Flipping the less than to greater than works for the real input but not
the test. So I implemented a janky switch to flip the comparison if
you've done more than 1,000 iterations.

I don't know why this doesn't work for both cases, but it gets the
answer 🙃