2022 day 24
There and back [and there] again.

Getting the blizzard moving was a little chore, but then I read the
instructions again and realised their positions were time based. Then a
BFS to find our target... which didn't work because you can stand still!
AND adding a wall at the start/end stops the search just going on a
jolly around the outside of the wall

Overall fun