2023 day 11
Manhattan distance! Recognised it quite quickly (good for so early!) but
I didn't make a sparse grid for part 1. So one small refactor later and
we are off!

Making the non-sparse grid was tricky, but the current solution is 👌
It could be even simpler and not use a dict, but 🤷‍♂️